ai/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sentence-transformers==2.5.1
qdrant-client==1.11.0
python-dotenv==1.0.1
pydantic==2.6.1
